+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### SDFormat 8.X.X (201X-XX-XX)
 
+1. Added Sensor DOM `==` operator.
+   * [Pull request 529](https://bitbucket.org/osrf/sdformat/pull-requests/529)
+
 1. Added AirPressure SDF DOM
    * [Pull request 528](https://bitbucket.org/osrf/sdformat/pull-requests/528)